Extract javascript code to a separate file.
